Make TimeFrameSet.Builder into an interface, move the implementation into TimeFrames
